requirements.txt
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.1
Flask-WTF==0.15.1
google-api-core==2.0.0
google-auth==2.0.1
google-cloud-bigquery==2.24.1
google-cloud-core==2.0.0
google-crc32c==1.1.2
google-resumable-media==2.0.0
googleapis-common-protos==1.53.0
grpcio==1.39.0
idna==3.2
importlib-metadata==4.6.4
itsdangerous==2.0.1
Jinja2==3.0.1
mailjet-rest==1.3.4
MarkupSafe==2.0.1
packaging==21.0
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dotenv==0.19.0
requests==2.26.0
rsa==4.7.2
six==1.16.0
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==2.0.1
WTForms==2.3.3
zipp==3.5.0
